[Workplace Search] Refactor nav constants (#82646)
* Move NAV constants to top level

This was only needed locally in Groups but we can to store all  nav constants in the global constants file

* Extract remaining nav constants

* Fix links to NAV in routers

* Use constant for path
