+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/components/layout/nav.tsx
@@ -4,7 +4,6 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import React from 'react';
-import { i18n } from '@kbn/i18n';
 
 import { EuiSpacer } from '@elastic/eui';
 
@@ -13,7 +12,7 @@ import { getWorkplaceSearchUrl } from '../../../shared/enterprise_search_url';
 import { SideNav, SideNavLink } from '../../../shared/layout';
 
 import { GroupSubNav } from '../../views/groups/components/group_sub_nav';
-import { NAV } from '../../views/groups/constants';
+import { NAV } from '../../constants';
 
 import {
   ORG_SOURCES_PATH,
@@ -29,38 +28,26 @@ export const WorkplaceSearchNav: React.FC = () => {
   return (
     <SideNav product={WORKPLACE_SEARCH_PLUGIN}>
       <SideNavLink to="/" isRoot>
-        {i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.overview', {
-          defaultMessage: 'Overview',
-        })}
+        {NAV.OVERVIEW}
       </SideNavLink>
       <SideNavLink isExternal to={getWorkplaceSearchUrl(ORG_SOURCES_PATH)}>
-        {i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.sources', {
-          defaultMessage: 'Sources',
-        })}
+        {NAV.SOURCES}
       </SideNavLink>
       <SideNavLink to={GROUPS_PATH} subNav={<GroupSubNav />}>
         {NAV.GROUPS}
       </SideNavLink>
       <SideNavLink isExternal to={getWorkplaceSearchUrl(`#${ROLE_MAPPINGS_PATH}`)}>
-        {i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.roleMappings', {
-          defaultMessage: 'Role Mappings',
-        })}
+        {NAV.ROLE_MAPPINGS}
       </SideNavLink>
       <SideNavLink isExternal to={getWorkplaceSearchUrl(`#${SECURITY_PATH}`)}>
-        {i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.security', {
-          defaultMessage: 'Security',
-        })}
+        {NAV.SECURITY}
       </SideNavLink>
       <SideNavLink isExternal to={getWorkplaceSearchUrl(ORG_SETTINGS_PATH)}>
-        {i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.settings', {
-          defaultMessage: 'Settings',
-        })}
+        {NAV.SETTINGS}
       </SideNavLink>
       <EuiSpacer />
       <SideNavLink isExternal to={getWorkplaceSearchUrl(`#${SOURCES_PATH}`)}>
-        {i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.personalDashboard', {
-          defaultMessage: 'View my personal dashboard',
-        })}
+        {NAV.PERSONAL_DASHBOARD}
       </SideNavLink>
     </SideNav>
   );
