--- x-pack/plugins/enterprise_search/public/applications/workplace_search/views/groups/constants.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { i18n } from '@kbn/i18n';
-
-export const NAV = {
-  GROUPS: i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.groups', {
-    defaultMessage: 'Groups',
-  }),
-  OVERVIEW: i18n.translate('xpack.enterpriseSearch.workplaceSearch.nav.groups.groupOverview', {
-    defaultMessage: 'Overview',
-  }),
-  SOURCE_PRIORITIZATION: i18n.translate(
-    'xpack.enterpriseSearch.workplaceSearch.nav.groups.sourcePrioritization',
-    { defaultMessage: 'Source Prioritization' }
-  ),
-};
